% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boxhist.R
\name{print.boxhist}
\alias{print.boxhist}
\title{print.boxhist}
\usage{
\method{print}{boxhist}(x, ...)
}
\arguments{
\item{x}{a list}

\item{which}{which variable to select}
}
\value{
a boxplot and histogram
}
\description{
Prints both the boxplot and histogram for the given data frame
}
\examples{
\dontrun{
 l <- boxhist(ddt)
 print(l)}
}
